% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{act_rank}
\alias{act_rank}
\title{Conduct factorization rank selection based upon ACT}
\usage{
act_rank(x, family = gaussian(), weights = 1, offset = zeros(x))
}
\arguments{
\item{x}{Input matrix to be factorized.}

\item{family}{Family object to specify deviance loss.}

\item{weights}{Entrywise weight.}

\item{offset}{Entrywise offset.}
}
\description{
Conduct factorization rank selection based upon ACT
}
