% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_center}
\alias{dmf_center}
\title{Center \code{dmf}.}
\usage{
dmf_center(lv, x0 = rep(1, nrow(lv$L)), reorder = TRUE)
}
\arguments{
\item{lv}{DMF structure to be centered.}

\item{x0}{Intercept column.}

\item{reorder}{Reorder columns according to norm of Lambda?}
}
\value{
Centered DMF structure with \code{center} attribute.
}
\description{
Center \code{dmf}.
}
