% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_extend}
\alias{dmf_extend}
\title{Extend rank from q to an arbitrary rank.}
\usage{
dmf_extend(
  x,
  dx,
  extend_rank = ncol(dx$L) + 1,
  adjust_svd = FALSE,
  refine = FALSE,
  control = glm.control(epsilon = 1e-06, maxit = 100)
)
}
\arguments{
\item{x}{The matrix that was partially decomposed}

\item{dx}{DMF structure for x of rank q}

\item{extend_rank}{new rank, defaults to q + 1}

\item{adjust_svd}{Should the DMF be orthogonalized after increasing rank?}

\item{refine}{refine new DMF with another iteration?}

\item{control}{algorithm control parameters (see \code{glm.control}).}
}
\value{
Extended DMF structure.
}
\description{
Extend rank from q to an arbitrary rank.
}
