% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_fitted}
\alias{dmf_fitted}
\title{Fitted values from a DMF.}
\usage{
dmf_fitted(lv, rank = ncol(lv$L), subset = seq_len(nrow(lv$L)), linear = FALSE)
}
\arguments{
\item{lv}{DMF structure}

\item{rank}{rank for the fitted values, that is, number of columns used in the factors}

\item{subset}{row subset for the values}

\item{linear}{return linear predictor?}
}
\value{
Fitted or predicted values.
}
\description{
Fitted values from a DMF.
}
