% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_identify}
\alias{dmf_identify}
\title{Conduct force-sign SVD given eta and desired rank q}
\usage{
dmf_identify(eta, q)
}
\arguments{
\item{eta}{Input matrix to conduct force-sign svd.}

\item{q}{designed rank of the force-sign svd.}
}
\description{
Conduct force-sign SVD given eta and desired rank q
}
