% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_multinom}
\alias{dmf_multinom}
\title{Perform a deviance matrix factorization, for a multinomial tensor.}
\usage{
dmf_multinom(
  x,
  rank = dim(x)[3],
  control = glm.control(epsilon = 1e-06, maxit = 100)
)
}
\arguments{
\item{x}{Input tensor to be factorized, of dimensions (n x k) x p.}

\item{rank}{Decomposition rank.}

\item{control}{Algorithm control parameters (see \code{glm.control}).}
}
\value{
DMF structure, where L is (n x k) x rank and V is p x rank.
}
\description{
Perform a deviance matrix factorization, for a multinomial tensor.
}
