% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_rank1update}
\alias{dmf_rank1update}
\title{Increases the rank of a DMF.}
\usage{
dmf_rank1update(dx, x, adjust_svd = FALSE, maxit = 100, refine = TRUE, ...)
}
\arguments{
\item{dx}{DMF structure for x of rank q.}

\item{x}{The matrix that was partially decomposed.}

\item{adjust_svd}{Should the DMF be orthogonalized after increasing rank?}

\item{maxit}{Maximum number of iterations for rank-one offset DMF.}

\item{...}{Other parameters?}
}
\value{
DMF structure for x of rank q + 1.
}
\description{
Increases the rank of a DMF.
}
