% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_residuals}
\alias{dmf_residuals}
\title{Residual values from a DMF.}
\usage{
dmf_residuals(
  x,
  lv,
  rank = ncol(lv$L),
  subset = seq_len(nrow(lv$L)),
  pearson = FALSE
)
}
\arguments{
\item{x}{data matrix}

\item{lv}{DMF structure, dim-consistent with data \code{x}}

\item{rank}{rank for the fitted values, that is, number of columns used in the factors}

\item{subset}{row subset for the values}

\item{pearson}{Pearson residuals?}
}
\value{
Residual matrix.
}
\description{
Residual values from a DMF.
}
