% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_sample}
\alias{dmf_sample}
\title{Sample data from a dmf linear predictor.}
\usage{
dmf_sample(family, eta, phi = 1, glm_weights = 1)
}
\arguments{
\item{family}{GLM family object specifying sampling distribution}

\item{eta}{linear predictor}

\item{phi}{dispersion parameter, defaults to unit}

\item{weights}{family weights}
}
\description{
Sample data from a dmf linear predictor.
}
