% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_sign}
\alias{dmf_sign}
\title{Normalize sign of \code{dmf}: the first row of Lambda in \code{dmf} is non-negative.}
\usage{
dmf_sign(lv)
}
\arguments{
\item{lv}{DMF structure to be sign-normalized.}
}
\value{
Sign-normalized DMF structure.
}
\description{
Normalize sign of \code{dmf}: the first row of Lambda in \code{dmf} is non-negative.
}
