% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_symm}
\alias{dmf_symm}
\title{Perform a symmetric deviance matrix factorization, (L * V' + V * L') / 2.}
\usage{
dmf_symm(
  x,
  family = gaussian(),
  rank = ncol(x),
  weights = 1,
  offset = zeros(x),
  maxit = 100,
  epsilon = 1e-06,
  trace = TRUE
)
}
\arguments{
\item{x}{Input matrix to be factorized.}

\item{family}{Family object to specify deviance loss.}

\item{rank}{Decomposition rank.}

\item{weights}{Entrywise weight.}

\item{offset}{Entrywise offset.}

\item{control}{Algorithm control parameters (see \code{glm.control}).}
}
\value{
DMF structure.
}
\description{
Perform a symmetric deviance matrix factorization, (L * V' + V * L') / 2.
}
