% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{family_test}
\alias{family_test}
\title{Perform a generalized Hosmer–Lemeshow test , (L * V' + V * L') / 2.}
\usage{
family_test(x, lv, G, weights = 1, offset = zeros(x), dispersion = NULL)
}
\arguments{
\item{x}{Input matrix to be factorized.}

\item{lv}{DMF structure to be tested.}

\item{G}{number of groups for eta to be partitioned}

\item{weights}{Entrywise weight.}

\item{offset}{Entrywise offset.}

\item{dispersion}{If cutomized dispersion is provided}
}
\description{
Perform a generalized Hosmer–Lemeshow test , (L * V' + V * L') / 2.
}
