% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{onatski_rank}
\alias{onatski_rank}
\title{Conduct factorization rank selection based upon maximum eigenvalue gap
Onatski(2010). Algorithm first fit dmf with rank = q_max then conduct eigen
gap search on the covariance of fitted eta}
\usage{
onatski_rank(
  x,
  family,
  q_max,
  weights = 1,
  offset = zeros(x),
  max_iter = 10,
  fit_full = FALSE
)
}
\arguments{
\item{x}{Input matrix to be factorized.}

\item{family}{Family object to specify deviance loss.}

\item{q_max}{Maximum rank eta.}

\item{weights}{Entrywise weight.}

\item{offset}{Entrywise offset.}

\item{max_iter}{Maximum iteration.}

\item{fit_full}{Fit eta with full rank matrix? Always recommended if computation speed is permitted}
}
\description{
Conduct factorization rank selection based upon maximum eigenvalue gap
Onatski(2010). Algorithm first fit dmf with rank = q_max then conduct eigen
gap search on the covariance of fitted eta
}
