% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_extend}
\alias{dmf_extend}
\title{extend rank from q to an arbitrary rank \verb{\\tilde\{q\}} through offset = eta_q}
\usage{
dmf_extend(
  x,
  dx,
  extend_rank = rank + 1,
  adjust_svd = FALSE,
  refine = FALSE,
  control = glm.control(epsilon = 1e-06, maxit = 100)
)
}
\arguments{
\item{x}{Input matrix to be factorized.}

\item{dx}{dmf rank q fit result}

\item{adjust_svd}{Todo:fixme.}

\item{refine}{boolen to indicate whether or not to refine rank q + 1 fit \verb{\\Lambda} V^T}

\item{control}{Algorithm control parameters (see \code{glm.control}).}
}
\value{
Extended DMF structure.
}
\description{
extend rank from q to an arbitrary rank \verb{\\tilde\{q\}} through offset = eta_q
}
