% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_rank1update}
\alias{dmf_rank1update}
\title{Increases the rank of a DMF.}
\usage{
dmf_rank1update(dx, x, ...)
}
\arguments{
\item{dx}{DMF structure for x of rank q.}

\item{x}{The matrix that was partially decomposed.}
}
\value{
DMF structure for x of rank q + 1.
}
\description{
Increases the rank of a DMF.
}
