% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_rank1update}
\alias{dmf_rank1update}
\title{Increases the rank of a DMF.}
\usage{
dmf_rank1update(
  x,
  dx,
  adjust_svd = FALSE,
  refine = TRUE,
  control = glm.control(epsilon = 1e-06, maxit = 100),
  ...
)
}
\arguments{
\item{x}{The matrix that was partially decomposed.}

\item{dx}{DMF structure for x of rank q.}

\item{adjust_svd}{Should the DMF be orthogonalized after increasing rank?}

\item{refine}{refine new DMF with another iteration?}

\item{control}{algorithm control parameters (see \code{glm.control}).}

\item{...}{other parameters passed to dmf calls}

\item{maxit}{Maximum number of iterations for rank-one offset DMF.}
}
\value{
DMF structure for x of rank q + 1.
}
\description{
Increases the rank of a DMF.
}
