% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmf.R
\name{dmf_rank1update}
\alias{dmf_rank1update}
\title{Increases the rank of a DMF.}
\usage{
dmf_rank1update(
  dx,
  x,
  adjust_svd = FALSE,
  refine = TRUE,
  control = glm.control(epsilon = 1e-06, maxit = 100),
  ...
)

dmf_rank1update(
  dx,
  x,
  adjust_svd = FALSE,
  refine = TRUE,
  control = glm.control(epsilon = 1e-06, maxit = 100),
  ...
)
}
\arguments{
\item{dx}{dmf rank q fit result}

\item{x}{Input matrix to be factorized.}

\item{adjust_svd}{Todo:fixme.}

\item{refine}{boolen to indicate whether or not to refine rank q + 1 fit \verb{\\Lambda} V^T}

\item{control}{Algorithm control parameters (see \code{glm.control}).}

\item{...}{Other parameters?}

\item{maxit}{Maximum number of iterations for rank-one offset DMF.}
}
\value{
DMF structure for x of rank q + 1.

Extended DMF structure.
}
\description{
Increases the rank of a DMF.

extend rank from q to an arbitrary rank \verb{\\tilde\{q\}} through offset = eta_q
}
